split sender.c to multiple files
